+++ client/cl_fx.c
@@ -1214,13 +1214,13 @@ void CL_BigTeleportParticles (vec3_t org)
 
 		angle = Q_PI*2*(rand()&1023)/1023.0;
 		dist = rand()&31;
-		p->org[0] = org[0] + cos(angle)*dist;
-		p->vel[0] = cos(angle)*(70+(rand()&63));
-		p->accel[0] = -cos(angle)*100;
+		p->org[0] = org[0] + cosf(angle)*dist;
+		p->vel[0] = cosf(angle)*(70+(rand()&63));
+		p->accel[0] = -cosf(angle)*100;
 
-		p->org[1] = org[1] + sin(angle)*dist;
-		p->vel[1] = sin(angle)*(70+(rand()&63));
-		p->accel[1] = -sin(angle)*100;
+		p->org[1] = org[1] + sinf(angle)*dist;
+		p->vel[1] = sinf(angle)*(70+(rand()&63));
+		p->accel[1] = -sinf(angle)*100;
 
 		p->org[2] = org[2] + 8 + (rand()%90);
 		p->vel[2] = -100 + (rand()&31);
@@ -1644,8 +1644,8 @@ void CL_RailTrail (vec3_t start, vec3_t end)
 		VectorClear (p->accel);
 
 		d = i * 0.1f;
-		c = cos(d);
-		s = sin(d);
+		c = cosf(d);
+		s = sinf(d);
 
 		VectorScale (right, c, dir);
 		VectorMA (dir, s, up, dir);
@@ -1841,14 +1841,14 @@ void CL_FlyParticles (vec3_t origin, int count)
 	for (i=0 ; i<count ; i+=2)
 	{
 		angle = ltime * avelocities[i][0];
-		sy = sin(angle);
-		cy = cos(angle);
+		sy = sinf(angle);
+		cy = cosf(angle);
 		angle = ltime * avelocities[i][1];
-		sp = sin(angle);
-		cp = cos(angle);
+		sp = sinf(angle);
+		cp = cosf(angle);
 		angle = ltime * avelocities[i][2];
-		sr = sin(angle);
-		cr = cos(angle);
+		sr = sinf(angle);
+		cr = cosf(angle);
 	
 		forward[0] = cp*cy;
 		forward[1] = cp*sy;
@@ -1863,7 +1863,7 @@ void CL_FlyParticles (vec3_t origin, int count)
 
 		p->time = cl.time;
 
-		dist = sin(ltime + i)*64;
+		dist = sinf(ltime + i)*64;
 		p->org[0] = origin[0] + bytedirs[i][0]*dist + forward[0]*BEAMLENGTH;
 		p->org[1] = origin[1] + bytedirs[i][1]*dist + forward[1]*BEAMLENGTH;
 		p->org[2] = origin[2] + bytedirs[i][2]*dist + forward[2]*BEAMLENGTH;
@@ -1940,14 +1940,14 @@ void CL_BfgParticles (entity_t *ent)
 	for (i=0 ; i<NUMVERTEXNORMALS ; i++)
 	{
 		angle = ltime * avelocities[i][0];
-		sy = sin(angle);
-		cy = cos(angle);
+		sy = sinf(angle);
+		cy = cosf(angle);
 		angle = ltime * avelocities[i][1];
-		sp = sin(angle);
-		cp = cos(angle);
+		sp = sinf(angle);
+		cp = cosf(angle);
 		angle = ltime * avelocities[i][2];
-		sr = sin(angle);
-		cr = cos(angle);
+		sr = sinf(angle);
+		cr = cosf(angle);
 	
 		forward[0] = cp*cy;
 		forward[1] = cp*sy;
@@ -1962,7 +1962,7 @@ void CL_BfgParticles (entity_t *ent)
 
 		p->time = cl.time;
 
-		dist = sin(ltime + i)*64;
+		dist = sinf(ltime + i)*64;
 		p->org[0] = ent->origin[0] + bytedirs[i][0]*dist + forward[0]*BEAMLENGTH;
 		p->org[1] = ent->origin[1] + bytedirs[i][1]*dist + forward[1]*BEAMLENGTH;
 		p->org[2] = ent->origin[2] + bytedirs[i][2]*dist + forward[2]*BEAMLENGTH;
@@ -1972,7 +1972,7 @@ void CL_BfgParticles (entity_t *ent)
 
 		VectorSubtract (p->org, ent->origin, v);
 		dist = VectorLength(v) / 90.0;
-		p->color = floor (0xd0 + dist * 7);
+		p->color = floorf (0xd0 + dist * 7);
 		p->colorvel = 0;
 
 		p->alpha = 1.0f - dist;
